Better Python support for failing constructors.

+++ backends/csharp/tests/output_safe/Test.cs
@@ -80,7 +80,7 @@ namespace interop_test
         [Fact]
         public void pattern_service_generated()
         {
-            var x = new SimpleService(123);
+            var x = new SimpleService((uint) 123);
             var b = new byte[] { 1, 2, 3 } ;
             
             x.MethodMutSelfFfiError(b);
